Deprecation: uint8_array.toString()
Gjs now returns Uint8Array instead of the custom ByteArray form
introspected functions. This means that in the future array.toString()
on a Uint8Array will return a string with the decimal digits of each
byte joined with commas.

To get the old behavior of decoding the to decode the byte array into a
string with default encoding UTF-8 ByteArray.toString() needs to be
used.